\name{extract.from.template}
\alias{extract.from.template}
\title{Extract part of a template}
\usage{
  extract.from.template(template, int)
}
\arguments{
  \item{template}{A recursive template}

  \item{int}{a signle integer or an interval of the form
  \code{c(n,m)}.}
}
\description{
  Extract part of a template
}

