\name{real.names}
\alias{real.names}
\title{Get real names of an exeres}
\usage{
  real.names(x)
}
\arguments{
  \item{x}{an exeres object}
}
\value{
  A character vector of the stored names of \code{x} (as
  opposed to the aliases).
}
\description{
  Get real names of an exeres
}

